[css-masonry] Update Intrinsic Sizing Rows 003 FR Expected / Ref Test Case https://bugs.webkit.org/show_bug.cgi?id=282992 rdar://problem/139727829
Reviewed by NOBODY (OOPS!).

Updated the font to Ahem to tackle the computed width sizing issue of "3 3". Added colors to the different items to catch ordering issues.

* Grid 2

Correct row 1 and 4 heights.
Item 5 should be a height of 2ch.

* Grid 4

Remove hidden items as they were creating a new column, which caused a gap at the end.
Remove spans in items 2 and 3; they do not make any sense.

* Grid 5 - 9

Remove hidden items.
Remove spans on items 2 and 3.
Remove spans that were creating an extra column causing an extra gap, which messes with widths.

* LayoutTests/TestExpectations:
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/masonry/tentative/intrinsic-sizing/masonry-intrinsic-sizing-rows-003-fr-expected.html:
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/masonry/tentative/intrinsic-sizing/masonry-intrinsic-sizing-rows-003-fr-ref.html:
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/masonry/tentative/intrinsic-sizing/masonry-intrinsic-sizing-rows-003-fr.html: